Fixes made based on release validator

+++ LifeCountDown/YLProgressBar.m
@@ -342,7 +342,7 @@ const NSTimeInterval YLProgressBarProgressTime = 0.25f; // s
         CGContextClip(context);
         
         CFArrayRef colorRefs = (__bridge CFArrayRef)_colors;
-        int colorCount = [_colors count];
+        NSInteger colorCount = [_colors count];
         
         float delta = 1.0f / [_colors count];
         float semi_delta = delta / 2.0f;
@@ -367,10 +367,10 @@ const NSTimeInterval YLProgressBarProgressTime = 0.25f; // s
     CGContextSaveGState(context); {
         UIBezierPath *allStripes = [UIBezierPath bezierPath];
         
-        int start = -_stripesWidth;
+        NSInteger start = -_stripesWidth;
         int end = rect.size.width / (2 * _stripesWidth) + (2 * _stripesWidth);
         CGFloat yOffset = (_type == YLProgressBarTypeRounded) ? YLProgressBarSizeInset : 0;
-        for (int i = start; i <= end; i++) {
+        for (NSInteger i = start; i <= end; i++) {
             UIBezierPath *stripe = [self stripeWithOrigin:CGPointMake(i * 2 * _stripesWidth + _stripesOffset, yOffset)
                                                    bounds:rect
                                               orientation:_stripesOrientation];
